+++ homeassistant/components/vacuum/xiaomi_miio.py
@@ -2,7 +2,7 @@
 Support for the Xiaomi vacuum cleaner robot.
 
 For more details about this platform, please refer to the documentation
-https://home-assistant.io/components/vacuum.xiaomi/
+https://home-assistant.io/components/vacuum.xiaomi_miio/
 """
 import asyncio
 from functools import partial
@@ -27,7 +27,7 @@ _LOGGER = logging.getLogger(__name__)
 
 DEFAULT_NAME = 'Xiaomi Vacuum cleaner'
 ICON = 'mdi:google-circles-group'
-PLATFORM = 'xiaomi'
+PLATFORM = 'xiaomi_miio'
 
 PLATFORM_SCHEMA = PLATFORM_SCHEMA.extend({
     vol.Required(CONF_HOST): cv.string,
