+++ tests/components/vacuum/test_xiaomi_miio.py
@@ -11,7 +11,7 @@ from homeassistant.components.vacuum import (
     SERVICE_CLEAN_SPOT, SERVICE_LOCATE, SERVICE_RETURN_TO_BASE,
     SERVICE_SEND_COMMAND, SERVICE_SET_FAN_SPEED, SERVICE_START_PAUSE,
     SERVICE_STOP, SERVICE_TOGGLE, SERVICE_TURN_OFF, SERVICE_TURN_ON)
-from homeassistant.components.vacuum.xiaomi import (
+from homeassistant.components.vacuum.xiaomi_miio import (
     ATTR_CLEANED_AREA, ATTR_CLEANING_TIME, ATTR_DO_NOT_DISTURB, ATTR_ERROR,
     CONF_HOST, CONF_NAME, CONF_TOKEN, PLATFORM,
     SERVICE_MOVE_REMOTE_CONTROL, SERVICE_MOVE_REMOTE_CONTROL_STEP,
